moved Config class into models/config.py
